Fix kafka produce consume test
The error in the produce consume test was shwallowed by rpk.
```
unable to create topics [produce.consume.test]: unable to dial: dial tcp 127.0.0.1:9092: connect: connection refused
/bin/bash: line 1: exit: too many arguments
```

`REDPANDA_BROKERS` was added to the pod template spec as rpk in tls test
need to have fully qualified name as it is defined in self signed certificate.

As the definition of the produce consume test changed from pod to job the
topic needs to be uqnique per job run, that is why pod name is used to
randomise the topic name.

Produce and consume test can now failed up to 2 times, before it is considered
as failed.