Console: Set same key for login-oidc-client-secret in Secret and Deployment
Currently, in the Helm chart the name for the login-oidc-client-secret key in the Secret does not match the key reference in the Deployment. In the Deployment the key login-oidc-client-secret is referenced, but in the Secret it is named login-generic-oidc-client-secret. I would drop the "generic-" part in the Secret, because in the values file it is configured under secret.login.oidc.clientSecret, also without "generic".

Key in the Deployment: https://github.com/redpanda-data/helm-charts/blob/main/charts/console/templates/deployment.yaml#L207