redpanda: don't set `tlsConfig` in ServiceMonitor if `tls.enabled` is false
Previously, if `tls.enabled` was false but monitoring was enabled
(`monitoring.enabled=true`) the `ServiceMonitor` object would
incorrectly render a `tlsConfig` and set `scheme` to `https`.

This was due to the default value of `monitoring.tlsConfig={}` being
evaluated as truthy.

This commit removes the `monitoring.tlsConfig` check entirely as there
are further checks within the stanza. To mitigate any confusion an note
as been added to the values.yml file.

Fixes #573