Add fixed sleep for test to wait for connector to finish replicating data
The rpk topic consume could fail for the first few times as kafka connect needs
to spawn the task and copy one message from the source topic. To solve this race condition
the retry should be implemented in bash for rpk topic consume or other mechanism that
can confirm source connectors started its execution. As a fast fix fixed 30 second fix is added.