replace task installs with nix
Previously, taskfile was being used to manage our development
environment and a variety of other techniques for CI. Taskfile is poorly
suited for this job as is evident by the difficulty leveraging it within
CI.

This commit begins the migration towards a nix based development
environment. Doing so allows us to keep our development and CI
environments the same. It also makes taskfiles easier to write as it can
be assumed that all binaries will be available via `$PATH`.

Individual developers may still opt to install the required tooling via
other means. See CONTRIBUTING.md for more details.