redpanda: Add timeouts to curl exec probes
Thanks to digging from Rafal, we've discovered that exec probes in
kubernetes implicitly get an additional 2 minute timeout [1]. This
results in abnormally long startup times for redpanda as each Pod's
startup probe can hang for up to 2 minutes.

This commit adds a timeout of 5 seconds to each exec using curl via the
`-m` flag which prevents the 2 minute hang time.

For posterity, 5 seconds was chosen arbitrarily.

[1]: https://github.com/kubernetes/kubernetes/blob/54f9807e1e84981b2053f4daf779f5ed19962144/pkg/kubelet/cri/remote/remote_runtime.go#L475-L478