Implement genpartial "partial" struct generator
This commit implements the `genpartial` go runnable for generating
"partial" variants of user defined structs.

By default, all fields of a struct will be included in the json/yaml
serialization. Developers may utilize the `omitempty` json tag to
exclude fields when they are the zero value of their respective type.
This however does not work with struct values unless the field is a
pointer to a struct with a values of `nil`. This behavior can lead to
various UX/DX issues, which you may have experienced if you've used the
go amazon SDK.

As we convert more of the redpanda chart into go, the exact shape of the
`Values` struct will become more important to avoid constant double
checking and second guessing. This struct should represent the union of
the default values and the user provided values, assuming they match the
JSON schema. This can be problematic when using the Values struct to define
test values as they'll be fully serialized and overwrite many of the
default values with go's zero values.

To work around this, we'll leverage the "partial" structs that are
generated by this new package.