Add chart-testing and chart-releaser to flake.nix
This commit adds the necessary nix files to install `chart-releaser`,
which is not available in nixpkgs, and joejuilian's fork of
`chart-testing`.

Doing so paves the way towards replacing unnecessary github actions
(anything that could be replaced with nix and a single shell command)
with nix to cut down on the duplication of dependencies.

This work was motivated by the sudden flakiness of installing github
actions. See also #1097.