Make chart_test.go idempotent
Due to the fact that in the charts/redpanda/Chart.yaml file the version
needs to be bumped every time when Redpanda helm chart release happen, the
charts/redpanda/Chart.lock in local environment will change over time. To make
snapshot testing idempotent the `helm.sh/chart` label is overwritten during
`helm templating`. During implementation of that feature it was discovered that
merging of the commonLabels variable is wrongly used in `FullLabels` and the
`statefulset-pod-labels` named template did not handle clash in user provided
`commonLabels` for the following labels:
* app.kubernetes.io/name
* app.kubernetes.io/instance
* app.kubernetes.io/component

As the development of the test suite revealed that current merge function,
that uses built-in maps package, does not work as the sprig equivalent, the
helmette package tests where extended with the same test case as with sprig.

The gotohelm test suite also changed, so that any implementation of function
that will be transpiled to helm template would not change the values in the
dot context. During the course of merge function implementation the
`values.commonLabels` was changed, so potentially other (next) usage of
`commonLabels` would be malformed.

https://github.com/Masterminds/sprig/blob/581758eb7d96ae4d113649668fa96acc74d46e7f/docs/dicts.md?plain=1#L125-L182

https://github.com/Masterminds/sprig/blob/581758eb7d96ae4d113649668fa96acc74d46e7f/dict_test.go#L142-L197