introduce pkg/valuesutil
This commit introduces `pkg/valuesutil` which contains a set of
utilities to work with helm values and jsonschemas. Most notably, it
includes the `Generate` function which can deterministically produce a
randomized output that is considered valid for a given json schema.

Future commits will utilize this utility to both fuzz and property test
the redpanda chart.

Examples of both fuzz testing and property testing with go's builtin
constructs are included in `fuzz_test.go`.