reduce cache busting for gotohelm
As genschema uses reflection, it needs to import the redpanda chart's
gocode which results in needing to rebuild gotohelm and genpartial much
more frequently than strictly is necessary which slows down development.

This commit splits the source regex filters for gotohelm, genpartial, and
genschema by each package such that changes to the redpanda chart don't
required gotohelm to be needlessly rebuilt. This should improve both CI
and local development iteration times.

As an added bonus, this commit also add `nix fmt` to CI.