gotohelm: add support for helm's `lookup` function
This commit adds support for the `lookup` function from helm. It does so by
injecting a Kubernetes API client into the `helmette.Dot` struct to avoid
relying on global variables. This is a divergence from helm itself but is
thoroughly documented as such.

Unlike helm's `lookup`, gotohelm returns `nil, false` in the case of not found
errors rather than an empty dictionary.