redpanda: move golden files into txtars
Previously, all golden files where kept as individual files in the `testdata`
directory. This would result in very noisy diffs on Github upon most changes to
the helm chart.

This commit adds functionality to run golden tests using a `txtar` archive
which is a plain text "tar" archive. This limits diffs to instead be to a small
set of very large files but preserves the ability to inspect diffs in plain
text.