operator: remove templates in favor of following kustomize
Previously there existed two sets of manifest for the operator. One in the
operator repo and one in the operator helm chart. Not only was the duplication
at risk of getting out of sync, there also existed a divergent feature set
between the two.

This commit removes as much of the operator helm chart as possible, instead
opting to follow the kustomize resources published by the operator itself.

This results in a few behavioral changes:
- {,Cluster}Role{,Bindings} names have changed and the total amount of
  permissions have been consolidated.
- `webhooks.enabled` now defaults to `true`. Disabling webhooks will likely
  result in a non-functional operator. This toggle will be removed in the
  future.
- `scope` is now deprecated. This poorly named field did not function the way
  it was documented and is no long possible to continue supporting due to the
  necessity of webhooks. It still exists and controls whether the operator runs
  in a namespaced mode or not BUT it does not affect the total set of RBAC
  objects created by this chart.