Add first function from helm chart engine
To handle Redpanda configuration in go code the `first` function needs to be
included to helmette as part of the sprig definition.

Negative numbers are recognized as `ast.UnaryExpr` with sub token. Now they can be transpiled.

P.S. The `[]byte` array can be represented only in Go, but not in template.