gotohelm: add numeric helper functions
This commit adds the `AsNumeric` and `AsIntegral` helper functions to the
`helmette` package to aid working with numeric values in gotohelm code.

A future commit will disallow type assertions and type checks of numeric values
in favor of these helpers as they are the closest approximation to keeping helm
and go behavior the same.