kminion: Add support for hostAliases (#1317)
* kminion: Add support for hostAliases

Some external private connectivity alternatives, such as Private Service Connect (PSC) (GCP),
PrivateLink (AWS) & Private Link (Azure) require incoming connections to use a different set of
hostnames and ports, which are configured as an additional set of advertised listeners in the
cluster.

These new advertised hostnames need not even exist in the VPC where Redpanda is deployed - it's
expected that client connections will come from a different VPC, likely in a different e.g. project
(GCP) or account (AWS). Therefore, the only way to measure SLA for them is to use the "path"
clients would use (i.e. the load balancer or ports configured for PSC/ PrivateLink/ Private Link),
and the only way to do that without creating additional DNS records in the Redpanda VPC (that would
only be used for monitoring, anyway), is to add a set of hostAliases to the kminion instance in
charge of monitoring it.

* Update kminion's README