gotohelm: Ban numeric type casting
Prior to this commit is was value to use type checks or assertions within
gotohelm code with numeric values. For example: `x.(int)`. Due to the numerous
limitations of go templates and helm, it's not feasible to fully support such
expressions as there's no type information that could be used to distinguish
between ints, int64s, and float64s.

This commit causes the transpiler to fail if it encounters such a statement and
instead points users to the recently added `AsNumeric` and `AsIntegral`
functions.