gotohelm: fix namespacing and function name overrides
Prior to this commit, there were subtle bugs with both package namespacing and
function name overriding that could result in template eval time errors.

This was due to directives only being parsed at the top loop of pkg.Syntax and
never being associated with objects later on. The ideal solution would be to
factor out a dedicated struct and cache for accessing directive information but
a temporary solution is being put in place for now.

The core issue was primarily only noticeable when calling bootstrap functions
from other bootstrap functions but a latent issue with name override also
existed. This commit adds a namespace and function naming cache that lazily
evaluate and cache their results that is now used in place of manually
constructing names.