gotohelm: add partial support for `resource.Quantity`
This commit adds partial support for Kubernetes' `resource.Quantity` type into
gotohelm. The supported methods are:
- `resource.MustParse`
- `resource.(*Quantity).Value`
- `resource.(*Quantity).String`
- `resource.(*Quantity).DeepCopy`

It is also possible to include a `resource.Quantity` type on a Values struct
and interact with it as you would any other `resource.Quantity` provided that
the value is a numeric or a valid `resource.Quantity` string.