*: add go.work file
Previously, this repository had a single `go.mod` file in the root and another
hidden one in gotohelm's testdata. This meant that writing any tests for
gotohelm would frequently require opening a new editor or LSP instance, which
is annoying at best.

This commit adds a `go.work` file to support a multi-module workspace. This
allows working on gotohelm tests and gotohelm within the same editor/LSP.