redpanda: convert secrets.yaml to go
This includes a fix to a long-standing bug: the curl arguments
for an admin listener that doesn't require a client cert named
the wrong file.

This replicates blank lines in generated shell scripts with high
fidelity to the helm originals, but that can be relaxed once
the conversion is complete.