redpanda: convert post-install-upgrade-job.yaml to go
This includes a new `helmette.MergeTo`, as follows:

We are largely dealing with types from three domains: the input
values, the desired k8s structures (both of these have relatively
sophisticated typing; although these types might in some cases be
roughly isomorphic but not identical) and the raw-json-structs of
Helm's templates.

Because we're coming from a templated world, the use of `merge` is
common to fold overrides onto defaults.

We provide a type-aware `MergeTo`, parameterised by the target type.
This permits go to be written in a type-safe fashion, but still permits
the kind of structure-merging common in code that begins life as the
translation of a Helm template. Under the hood, it has to round-trip
into `mergo`-compatible raw values, but we do our best to hide this
from the caller.