redpanda: move containers & volumes to go, update values schema accordingly
The following entries are now arrays of structured values rather than templated yaml strings:

- These are arrays of core/v1 `VolumeMount` objects:
-- values.statefulSet.initContainers.*.extraVolumeMounts
-- values.statefulSet.sideCars.*.extraVolumeMounts
-- values.statefulSet.extraVolumeMounts
- These are now arrays of core/v1 `Container` objects:
-- values.statefulSet.initContainers.extraInitContainers
- These are now arrays of core/v1 `Volume` objects:
-- values.statefulSet.extraVolumes

Updating helm values files for these is typically just a case of removing the `|-` from
the appropriate attribute definitions. Nested templating of these values [that is,
template code in values.yaml] is no longer supported. In practice,
this is unlikely to be required.

The intention is to move all this behaviour into the operator
ultimately, leaving the helm chart to simpy create a CRD. Under
those circumstances, templating of the CRD contents might still
be available *at the helm layer* but the operator will not
attempt to perform any further value expansion.