operator: bump app version to v2.1.26-24.1.9
This commit bumps the app version of the operator helm chart to v2.1.26-24.1.9.

It also fixes some merge skew which resulted in a redpanda chart file being out
of date.

It also comments out `{http,admin}.tls.cert` in the operator tests as this
triggers a bug in the redpanda chart which is blocking the operator release. A
future commit will correct this bug.