actions: run tests against nightly redpanda RCs
Previously, the nightly_redpanda_tip.yaml ran tests against the nightly build
of redpanda. This doesn't provide much value for the chart as nightlies can be
quite unstable.

This commit instead changes the action to run nightly tests against the most
recent redpanda RC build. This will provide an early and automatic signal to
the team if there are any breaking changes in upcoming versions of redpanda.