Handle empty string correctly using Empty function
In the sprig shim the empty string was not handled correctly. That's why
console tag in the Go vs Helm template unit test was failing. Go execution had
default value for console tag as empty string. The logic that took console tag
version from Chart appVersion was overwritten in the `containerImage` function.
In the end container did not have tag when Console chart was rendered.