gotohelm: match `metav1.Duration`'s marshaling
Prior to this commit the `MustParseDuration` would transpile to nothing. Any
string given to it would be passed through unchanged as the API server would
handle validation.

This is not acceptable as the go's `time.Duration` always adds includes minutes
and second in the stringified form. So `1h` would be rendered as `1h0m0s`. The
kube-apiserver accepts this as a valid duration but does not update it's form.
This can cause `.Generation` to be incremented when toggling between helm and
go as `0m0s` is added and removed from the apiserver.

This commit implements `time.ParseDuration` in gotohelm (with limitations) and
updates the transpilation of `MustParseDuration` to properly re-serialize the
provided duration.