remove gotohelm in favor of operator gotohelm
Prior to this commit `gotohelm` was duplicated in the `redpanda-operator` repo.

This commit removes `pkg/gotohelm` and instead depends on `pkg/gotohelm` from
the redpanda-operator repository. Due to the use of `replace` directives in the
go.mod of `redpanda-operator/gotohelm`, it's not currently possible to `go
install` it. As a temporary workaround, the `main.go` for gotohelm has been
preserved in this repository.

As gotohelm now requires go >= 1.23, this commit also syncs the `flake.lock`
for this repository with the redpanda-operator.