redpanda: fix out of range slice in go mode
This commit fixes a slice operation that could generate a panic when run as go
code if a string was less than 49 character in length. It has no effect on the
transpiled templates as `substr` appropriately handles out of bounds slices.

Ideally, this commit would modify `TestTemplate` to additionally run the go
chart as well however doing so surfaced other issues with `helmette.Tpl`. See
K8S-468.