Release from operator repo
Prior to this commit, the helm-charts repo hosted downstream copies of the
redpanda, console, connectors[^1] and operator charts.

To release these charts, we would manually copy the trees into this repo and
rely on chart-releaser.

This process was quite unpleasant as it lead to many surprises from the
workflows in this repo and caused a good amount of confusion for contributors.

This commit removes the mirrored charts and replaces the release process with a
new GHA that pulls release information from the redpanda-operator repository.
chart-releaser is leveraged to update the index.yaml in the gh-pages branch as
usual.

In addition to the removal, a new README.md has been added as the previous
symlink to the redpanda chart's README.md is now dead.

[^1]: While the connectors chart technically exists in the redpanda-operator,
it's been moved to community maintenance as of the v25.1.1 redpanda release.
It's source is being left in place as it's been removed from newer operator
releases.